chore: major robustness upgrade - thread-safe loading, CSV validation, network redundancy, and improved error logic

+++ backend/app/schemas/predict.py
@@ -5,9 +5,9 @@ from pydantic import BaseModel, Field
 
 class PredictRequest(BaseModel):
     """Prediction request body."""
-    jumlah_penjualan: int = Field(..., ge=0, examples=[150])
-    harga: int = Field(..., ge=0, examples=[50000])
-    diskon: int = Field(..., ge=0, le=100, examples=[10])
+    jumlah_penjualan: float = Field(..., ge=0, examples=[150])
+    harga: float = Field(..., ge=0, examples=[50000])
+    diskon: float = Field(..., ge=0, le=100, examples=[10])
 
 
 class PredictResponse(BaseModel):
